+++ src/pages/cart/Cart.tsx
@@ -2,8 +2,8 @@ import { useState, useEffect, useCallback } from 'react';
 import { apiGetCartById } from '@shared/api/commerce-tools/apiGetCartById';
 import { CartData } from '@shared/types/types';
 import { openDialog } from '@services/DialogService';
-import VoidCartArea from '@components/layout/cart/VoidCartArea.module';
-import CartInfo from '@components/layout/cart/CartInfo.module';
+import VoidCartArea from '@components/layout/cart/VoidCartArea';
+import CartInfo from '@components/layout/cart/CartInfo';
 import styles from '../../components/layout/cart/Cart.module.scss';
 import CartResult from '@components/layout/cart/CartResult';
 import { getDiscountDetails } from '@shared/api/commerce-tools/getDiscountDetails';
